changes for android shell

+++ img.sh
@@ -1,4 +1,2 @@
-#!/bin/bash
-
-dd if=/dev/block/bootdevice/by-name/boot of=boot.img
-dd if=/dev/block/bootdevice/by-name/dtbo of=dtbo-stock.img
+[ ! -f boot.img ] && dd if=/dev/block/bootdevice/by-name/boot of=boot.img
+[ ! -f dtbo-stock.img ] && dd if=/dev/block/bootdevice/by-name/dtbo of=dtbo-stock.img
